requirements.txt
Click==7.0
Flask==1.0.3
itsdangerous==1.1.0
Jinja2==2.10.1
kafka-python==1.4.6
MarkupSafe==1.1.1
Werkzeug==0.15.4
